requirements.txt
Flask==0.10.1
Flask-MongoKit==0.6
Flask-WTF==0.9.4
Jinja2==2.7.1
LogentriesLogger==0.2.1
MarkupSafe==0.18
WTForms==1.0.5
Werkzeug==0.9.4
gunicorn==18.0
itsdangerous==0.23
mongokit==0.9.1.1
pymongo==2.6.3
requests==2.0.1
